fixup! HACK: playbinpoolsrc: Force framerate of 30fps for video stream
When forcing the 30fps we could end up using the first buffer of the GOP
outputted by the decoder as first buffer, even if it was a long time in
the past. Ensuring that we only drop buffers in that videorate allows us
to properly clip buffers within the currently configured segment.