Add a new `segmentclipper` element
This element clips buffers based on the segment event it receives. It will drop buffers
that are outside the segment and will clip buffers that are partially outside the segment.
It expects all input buffer to have increasing PTS and will push EOS when the segment done.

When receiving an EOS event, it will push the last buffer it received if no buffer inside
the segment was received. This is useful in combination with the
GstVideoDecoder:drop-out-of-segment property, as it allow handling special cases
where the input stream has big gaps (in screen recording for example)
and seeking might lead to EOS without any buffer inside the segment, while what use would
expect is to have the frame before the gap to be displayed.